Move fake driver
